LayoutTests/fullscreen/enter-fullscreen-main-frame-window-open-iframe-expected.txt

To test manually, click the "Open window" button - the page should exit fullscreen mode.

EVENT(fullscreenchange)
TEST(document.fullscreenElement===document.body) OK
EVENT(fullscreenchange)
TEST(document.fullscreenElement===null) OK
END OF TEST
